pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "irce"
version = "0.1.0"
description = "Intelligent RecoveryControl Environment scaffold built on OpenEnv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "requests>=2.32.0",
    "typing-extensions>=4.12.0",
    "openai>=1.52.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
server = "irce.server.main:main"
irce-server = "irce.server.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
